add breadcrumb and item info on rent_records/review page

+++ app/helpers/items_helper.rb
@@ -1,9 +1,9 @@
 module ItemsHelper
 
   def render_collect_item_link(item)
-    if current_user && current_user.is_collected?(item)
+    if current_user && current_user.is_collected?(item) && collect_rent_display?
        link_to("取消收藏", uncollect_item_path(item), method: :delete, class: "btn btn-default")
-    else
+    elsif collect_rent_display?
        link_to("收藏", collect_item_path(item), method: :post, class: "btn btn-info")
     end
   end
@@ -15,6 +15,14 @@ module ItemsHelper
     when "subcategories"
       breadcrumb :subcategory, @subcategory
     end
+  end
+
+  def rent_record_button_display?(item)
+    current_user != item.lender && collect_rent_display?
+  end
+
+  def collect_rent_display?
+    !["review", "index"].include?(params[:action])
   end
 
 end
